requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.68.1
setuptools==58.1.0
respx==0.20.1
uvicorn==0.15.0
pytest-asyncio==0.21.0
tritonclient[http]==2.34.0
numpy==1.23.5
requests==2.31.0
gevent==22.10.2
geventhttpclient==2.0.2
aiohttp==3.8.4
opencv-python==4.7.0.72
torch==2.0.0
torchvision==0.15.0
python-multipart==0.0.6
